+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-02 09:59:14
-// Commit: feat(ui): add bi-directional event bus handler for ui
+// Updated at: 2026-04-02 12:55:15
+// Commit: docs(ui): update readme and API contracts for ui
 
 class uiComponentState {
-  final String commitRef = '43';
-  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
-  final DateTime timestamp = DateTime.parse('2026-04-02T09:59:14+05:00');
+  final String commitRef = '54';
+  final String lastMessage = 'docs(ui): update readme and API contracts for ui';
+  final DateTime timestamp = DateTime.parse('2026-04-02T12:55:15+05:00');
 }